move cli code to separated module
