+++ src/cli/commands/__init__.py
@@ -1,0 +1,1 @@
+from .get_started import get_started as get_started
